limit colortable index to uint8
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@3992 d0c010d8-bb82-477e-a075-112140230451

+++ src/IGarminTyp.cpp
@@ -996,7 +996,7 @@ bool IGarminTyp::decodeBppAndBytes(int ncolors, int w, int flags, int& bpp, int&
 }
 
 
-bool IGarminTyp::decodeColorTable(QDataStream& in, QImage& img, int ncolors, int maxcolor, bool hasAlpha)
+bool IGarminTyp::decodeColorTable(QDataStream& in, QImage& img, quint8 ncolors, int maxcolor, bool hasAlpha)
 {
 
 #ifdef QK_QT5_PORT
